Created send_messages() in client
This will in fact, match the freak of the recieve messsages function in client.py
I was on a bit of a wild goose chase but it seems that battleship.py formats messages for us,
hence the client and server need only exchange these strings.

I do not like this project.