Minimal Messaging "completed"
Client and Server will communicate to each other what current game state they are in.
This will be logged on the server log.
As inputs are already sent to the server and the game, our team found it counterintuitive to send extra messages that would
possibly interupt the flow of the game (ie the example the task sheet provided) thus we decided to sharing the game state
would be the most effective thing to message back and forth between the client and server.